Focus README on docker-compose; README fixups; reduce docker-compose dep
* README.md

Move Docker Compose instructions to the top of the types of setup, and
name it "Quick Start with Docker Compose".

Fix bug in instructions to append to /etc/default/telegraf in the bare
metal instructions (use sudo tee).

Added instructions for verifying installation by checking that
`nuodb_internal` and `nuodb` databases exist in influx.

Changed Kubernetes section to say "coming soon".

Fix header depth so that each of the Setup types are at the top level.

Fix whitespace.

* docker-compose.yml

Reduce dependency from 3.7 to 3.3, seems to work just fine and is
required for the versions of docker and docker-compose on Ubuntu.