Do not log to stdout in reload server
This suppresses the 'Error in plugin: metric parse error: expected field
at 1:11: " * Serving"' log message, which is due to Flask logging to
stdout, which Telegraf expects to be used to emit metrics data only.
This replaces the Flask development server with the WSGI reference
implementation server, which does not log anything to stdout.