Emit metrics from first SyncTrace message
The input plugin that emits metrics from SyncTrace messages filters data
if there is no change in number of stalls. This is problematic in the
tests, since there may be no change in number of stalls for a very long
time. The delta calculation is also a little weird, since it always
ignores the first sample, even if it actually contais a non-zero number
of stalls. This change emits metrics for the first SyncTrace message.